No auto ack and wait while leds updated
Change of strategy: when the individual ledstrip controllers send the 1.2MHz pulse trains to the ledstrips there is a lot of electrical noise. There is no point expecting to receive an acknowledgement at this point, nor is it worth sending a further update until the strip update is finished.

Therefore switch to non-auat-ack mode, use a single target address shared by all strips, and delay 10ms after sending an update before sending another.